+++ Samples~/Sample3/Scripts/MessageListView.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+using Kuuasema.DataBinding;
+
+
+public class MessageListViewModel : ListViewModel<Message,MessageDataModel,MessageView> {
+
+    public void Remove(MessageView messageView) {
+        this.DataModel.RemoveValue(messageView.DataModel.Value);
+    }
+
+}
+