+++ Samples~/Sample3/Scripts/MissionView.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+using Kuuasema.DataBinding;
+
+
+public class MissionView : MissionViewModel {
+    public void Claim() {
+        this.DataModel.Claim();
+    }
+}
+
+